simplify the logic of license
